move code that handles input event to input.cpp
